+++ customerService/src/main/java/customerService/customerServiceRepository/CustomerRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 import java.math.BigInteger;
 
 @Repository
-public interface CustomerRepository extends MongoRepository<Customer, BigInteger> {
+public interface CustomerRepository extends MongoRepository<Customer, Long> {
 }
